# Conflicts:
#	NaoTHSoccer/Source/Cognition/Modules/Behavior/XABSLBehaviorControl/Symbols/MotionSymbols.h
#	NaoTHSoccer/Source/Cognition/Modules/Behavior/XABSLBehaviorControl/agents.xabsl
#	NaoTHSoccer/Source/Representations/Motion/CollisionPercept.h
#	Utils/py/naoth/naoth/Representations_pb2.py

+++ Utils/CollectStick/startBrainwashing.sh
@@ -9,7 +9,7 @@
 check_for_errors() {
 	if [ "$?" -ne 0 ]
 	then
-		sudo -u nao /usr/bin/paplay /home/nao/naoqi/Media/error_while_collecting_logs.wav
+		su nao -c "/usr/bin/paplay /home/nao/naoqi/Media/error_while_collecting_logs.wav"
 		# if argument is available - write to systemlog
 		if [ ! -z "$1" ]
 		then
@@ -29,6 +29,7 @@ exec_cmd_and_return_or_default() {
 
 
 # set file vars
+currentDir=$(pwd)  # /media/brainwasher
 infoFile="/home/nao/Config/nao.info"
 errorFile="/home/nao/brainwasher.log"
 
@@ -55,24 +56,24 @@ current_compile_owner=$(exec_cmd_and_return_or_default 'grep "Owner" /var/log/me
 logger "Brainwasher:start $current_date, $current_nao, Player $current_nao_player"
 
 # play sound
-sudo -u nao /usr/bin/paplay /home/nao/naoqi/Media/usb_start.wav
+su nao -c "/usr/bin/paplay /home/nao/naoqi/Media/usb_start.wav"
 
 naoth stop
 
 # write previous volume to systemlog
-current_volume=$(sudo -u nao pactl list sinks | grep Volume | xargs)
+current_volume=$(su nao -c "pactl list sinks | grep Volume | xargs")
 logger "Brainwasher:$current_volume"
 
 # set volume to 88%
-sudo -u nao pactl set-sink-mute 0 false 2> $errorFile
-sudo -u nao pactl set-sink-volume 0 88% 2>> $errorFile
+su nao -c "/usr/bin/pactl set-sink-mute 0 false 2> $errorFile"
+su nao -c "/usr/bin/pactl set-sink-volume 0 88% 2>> $errorFile"
 logger -f $errorFile
 
 logger "Brainwasher:copy files"
 # create directory
 #dir_name=$current_date-$current_nao
 dir_name=${current_nao_player}_${current_nao_number}_${current_nao}_${current_date}
-target_path=/media/brainwasher/$dir_name
+target_path=$currentDir/$dir_name
 mkdir -p $target_path
 
 # copy info file of the nao
@@ -86,26 +87,38 @@ echo "$current_compile_time" >> $target_path/nao.info
 echo "$current_compile_owner" >> $target_path/nao.info
 
 # find log files and copy them to the created directory
-#find -L /tmp -type d -name media -prune -o -name "*.log" -exec cp {} /media/brainwasher/$current_date-$current_nao \;
+#find -L /tmp -type d -name media -prune -o -name "*.log" -exec cp {} $currentDir/$current_date-$current_nao \;
 # find log files, create MD5 hashes and copy everything to the created directory
 for f in $(find -L /tmp -type d -name media -prune -o -name "*.log")
 do
-	md5sum $f | sed -e "s/\/tmp\///g" > "$f.md5"
+	md5sum $f | sed -e "s#/tmp/##g" > "$f.md5"
+	cp "$f.md5" $f $target_path/
+	check_for_errors "Brainwasher:ERROR copying $f"
+done
+
+for f in $(find -L /dev/shm -type d -name media -prune -o -name "*.log")
+do
+	md5sum $f | sed -e "s#/dev/shm/##g" > "$f.md5"
 	cp "$f.md5" $f $target_path/
 	check_for_errors "Brainwasher:ERROR copying $f"
 done
 
 # copy the config directory
+logger "copy the config directory"
 cd /home/nao/naoqi
 zip -q -r -0 $target_path/config.zip Config
 cd -
 check_for_errors "Brainwasher:ERROR copying config"
 
 # copy logs of naoth binary (std::out/::err) and clear them afterwards
-cp "/var/log/naoth.log" "/var/log/naoth_err.log" $target_path/
-check_for_errors "Brainwasher:ERROR copying /var/log/naoth.log"
-> /var/log/naoth.log
-> /var/log/naoth_err.log
+if [ -f "/var/log/naoth.log" ]; then
+	cp "/var/log/naoth.log" "/var/log/naoth_err.log" $target_path/
+	check_for_errors "Brainwasher:ERROR copying /var/log/naoth.log"
+	> /var/log/naoth.log
+	> /var/log/naoth_err.log
+else
+	logger "naoth.log does not exist!"
+fi
 
 # find and copy trace dump files since boot and log errors
 find /home/nao -maxdepth 1 -type f -mmin -$current_boot_time -iname "trace.dump.*" -exec zip -q -0 $target_path/dumps.zip {} + 2> $errorFile
@@ -132,7 +145,7 @@ rm $errorFile
 sync
 
 # needed to play sound before starting naoth! otherwise the sound could get "lost" (no sound)
-sudo -u nao /usr/bin/paplay /home/nao/naoqi/Media/finished_collecting_logs.wav
+su nao  -c "/usr/bin/paplay /home/nao/naoqi/Media/finished_collecting_logs.wav"
 
 logger "Brainwasher:END, starting naoth"
